+++ MoreShipUpgrades/UpgradeComponents/Items/LategameItem.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace MoreShipUpgrades.UpgradeComponents.Items
+{
+    public abstract class LategameItem : GrabbableObject
+    {
+        public static void LoadItem()
+        {
+            throw new NotImplementedException();
+        }
+    }
+}